requirements.txt
# CS Tutor LLM Dependencies
# =========================

# Core ML
torch>=2.0.0
numpy>=1.24.0

# Tokenization
regex>=2023.0.0
sentencepiece>=0.1.99  # Optional: for LLaMA tokenizer compatibility

# Configuration
pyyaml>=6.0

# CLI
typer>=0.9.0
rich>=13.0.0

# Training (optional)
# wandb>=0.15.0  # For experiment tracking
# tensorboard>=2.14.0  # For visualization

# Quantization (optional)
# bitsandbytes>=0.41.0  # For 8-bit/4-bit on CUDA
# gguf>=0.1.0  # For GGUF export

# Development
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.0.0

